shader graph bug fix
Objects are trying to be loaded during a domain backup. This is not allowed as it will lead to undefined behaviour!

Memory issue due to missing 2D textures on the leaf and bush shader graphs. Needed to re-add the textures and restart the unity editor as it is in partial a memory issue